requirements.txt
aldryn-apphooks-config==0.6.0
asgiref==3.3.4
Babel==2.9.0
backport-collections==0.1
dj-database-url==0.5.0
Django==3.1.8
django-appconf==1.0.4
django-appdata==0.3.2
django-classy-tags==2.0.0
django-cms==3.8.0
django-filer==2.0.2
django-formtools==2.2
django-js-asset==1.2.2
django-meta==2.0.0
django-mptt==0.12.0
django-parler==2.2
django-phonenumber-field==1.3.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-sortedm2m==3.0.2
django-taggit==1.3.0
django-taggit-autosuggest==0.3.8
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-apphook-setup==0.4.1
djangocms-attributes-field==2.0.0
djangocms-blog==1.2.3
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy-thumbnails==2.7.1
html5lib==1.1
lxml==4.6.3
phonenumbers==8.12.21
phonenumberslite==8.12.21
Pillow==8.2.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
tzlocal==2.1
Unidecode==1.1.2
webencodings==0.5.1
YURL==1.0.0
djangorestframework~=3.12.4
pip~=21.1
setuptools~=41.2.0